# docs/swagger.yaml
definitions:
...
  handlers.UserUrlsResp:
    properties:
      original_url:
        type: string
      short_url:
        type: string
    type: object
info:
  contact: {}
  description: A URL shortening service with JWT cookie authentication.
  title: URL Shortener API
  version: "1.0"
paths:
...
  /api/user/urls:
    delete:
      consumes:
      - application/json
      description: Asynchronously deletes the given short URL aliases belonging to
        the authenticated user
      parameters:
      - description: List of aliases to delete
        in: body
        name: aliases
        required: true
        schema:
          items:
            type: string
          type: array
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad request or empty list
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Delete user's URLs
      tags:
      - user
    get:
      description: Returns all URLs shortened by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.UserUrlsResp'
            type: array
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Get user's URLs
      tags:
      - user
...
swagger: "2.0"